+++ src/lib/api.sh
@@ -33,7 +33,7 @@ add-lkc-binding(system, revision, lkc_directory, lkc_target=, lkc_output_directo
 
 # adds a kconfig model
 # implemented by library scripts (e.g., to read a kconfig model with kconfigreader, kclause, or configfix)
-add-kconfig-model(system, revision, kconfig_file, lkc_binding_file=, environment=) {
+add-kconfig-model(system, revision, kconfig_file, lkc_binding_file=, lkc_directory, lkc_target=, lkc_output_directory=, environment=) {
     :
 }