+++ src/systems/axtls.sh
@@ -5,6 +5,7 @@ AXTLS_URL=https://github.com/ekuiter/torte-axTLS
 
 add-axtls-kconfig-history(from=, to=) {
     add-system --system axtls --url "$AXTLS_URL"
+    add-hook-step configfix-pre-extraction-hook configfix-pre-extraction-hook-axtls
     for revision in $(git-tag-revisions axtls | exclude-revision @ | start-at-revision "$from" | stop-at-revision "$to"); do
         add-revision --system axtls --revision "$revision"
         add-kconfig \
@@ -13,4 +14,10 @@ add-axtls-kconfig-history(from=, to=) {
             --kconfig-file config/Config.in \
             --lkc-directory config/scripts/config
     done
+}
+
+configfix-pre-extraction-hook-axtls(system, revision) {
+    if [[ $system == axtls ]]; then
+        wrap-source-statements-in-double-quotes
+    fi
 }